+++ R/sim_pop.R
@@ -45,8 +45,7 @@ initSimPop=function(seed,bForce=FALSE){
 #' growthphase=sim_pop(NULL,params=params,cfg)
 sim_pop=function(tree,
                  params=list(),
-                 cfg=list(compartment=data.frame(val=c(0,1),rate=c(-1,1/120.0),popsize=c(1,1e4)),
-                          info=data.frame(val=c(0,1,1),fitness=c(-1,0,0.2))),
+                 cfg=getDefaultConfig(5e4,0.01),#list(compartment=data.frame(val=c(0,1),rate=c(-1,1/120.0),death_rate=c(0,0),popsize=c(1,1e4)),#    info=data.frame(val=c(0,1,1),fitness=c(-1,0,0.2))),
                  b_verbose=1, 
                  b_check=FALSE,
                  trajectory=NULL,  #tarjectory
@@ -54,14 +53,12 @@ sim_pop=function(tree,
                                        #will be picked.
 {
   toCombine=TRUE
-  ##browser()
   if(is.null(tree)){
     toCombine=FALSE
     tree=rtree(2)
     tree$edge.length=rep(0,2)
     tree$events=data.frame(value=0:1,driverid=c(0,0),node=1:2,ts=0.0,uid=0:1)
   }
-  #browser()
   # convert driversFitness into double to fit the input type for the C++ sim_pop() function
   if(is.null(driversFitness)){
     #driversFitness=-1.0
@@ -87,8 +84,6 @@ sim_pop=function(tree,
     max_initial_driverid=tree$maxDriverID
   }
   
-  
-  # print(tree$events)
   tree=standardiseTree(tree)  #populate the tree object with the #divisions (ndivs), the time (tBirth) and the maximum time (maxt)
   ndrivers=tree$ndrivers
   time=tree$tBirth
@@ -98,7 +93,8 @@ sim_pop=function(tree,
                      b_stop_at_pop_size=0,
                      maxt=tree$maxt,
                      driver_rate_per_cell_per_day=0,
-                     max_driver_count=-1
+                     max_driver_count=-1,
+                     nmigration=0
   )
   fields=names(defaultparams)
   if(length(setdiff(names(params),names(defaultparams)))>0){
@@ -140,45 +136,66 @@ sim_pop=function(tree,
   compartment=cfg$compartment
   compartment$popsize=ceiling(compartment$popsize)
   compartmentinfo=cfg$info
+  if(is.null(compartment$death_rate)){
+    compartment$death_rate=0.0
+  }
+  
   #browser()
   # set the trajectory columns to 0 if user is not running run_driver_process_sim()
   if(is.null(trajectory)){
-    trajectory_ts = 0; trajectory_pop = 0
-    trajectory_div_rate = 0; trajectory_size=0; trajectory_compartment=1
+    trajectory_ts = 0
+    trajectory_pop = 0
+    trajectory_div_rate = 0
+    trajectory_size=0
+    trajectory_compartment=1
+    trajectory_death_rate=0
+    trajectory_compartment2=1
+    trajectory_s_div_rate2=0
+    trajectory_a_div_rate2=0
     totalpop=max(sum(compartment$popsize),length(tree$tip.label))
-    MAX_SIZE=3*totalpop ##Allows for stochastic drift in population size (need to make this robust)
-    MAX_EVENTS=ceiling(2*params[["n_sim_days"]])
-  }else if(is.data.frame(trajectory) | is.data.table(trajectory)){
+    MAX_SIZE=3*totalpop +10000 ##Allows for stochastic drift in population size (need to make this robust)
+    MAX_EVENTS=ceiling(1000*params[["n_sim_days"]])
+  }else if(is.data.frame(trajectory) || is.data.table(trajectory)){
     trajectory$target_pop_size=ceiling(trajectory$target_pop_size)
     trajectory_ts = trajectory$ts
     trajectory_pop = trajectory$target_pop_size  
     trajectory_div_rate = trajectory$division_rate
+    trajectory_death_rate = trajectory$death_rate
     trajectory_size = nrow(trajectory)
     trajectory_compartment = trajectory$compartment
-    
+    trajectory_compartment2=trajectory$compartment
+    trajectory_s_div_rate2=rep(0,length(trajectory$compartment))
+    trajectory_a_div_rate2=rep(0,length(trajectory$compartment))
     trj=trajectory %>% group_by(compartment) %>% summarise(maxpop=max(target_pop_size))
     totalpop=sum(trj$maxpop) #max(trajectory_pop)   #max(sum(compartment$popsize),length(tree$tip.label))
     ##Include other compartments
     totalpop=totalpop+sum(compartment$popsize[which(!(compartment$val %in% trajectory$compartment))])
     totalpop=max(totalpop,length(tree$tip.label))
-    MAX_SIZE=ceiling(3*totalpop)+100 ##Allows for stochastic drift in population size (need to make this robust)
-    MAX_EVENTS=ceiling(2*max(trajectory_ts))
+    MAX_SIZE=ceiling(3*totalpop)+10000 ##Allows for stochastic drift in population size (need to make this robust)
+    MAX_EVENTS=ceiling(10*max(trajectory_ts))
   }
-  #browser()
-  # if(ndriver>1000){
-  #   stop("Too many drivers: edit CellSimulation::setCompartmentInfoRecursively in C++ to support more!")
-  # }
-  # totalpop=max(sum(compartment$popsize),length(tree$tip.label))
-  # MAX_SIZE=3*totalpop ##Allows for stochastic drift in population size (need to make this robust)
-  
-  # MAX_EVENTS=ceiling(2*params[["n_sim_days"]])
+  if(is.null(cfg$migrations) || nrow(cfg$migrations)==0){
+    nmigration=0
+    c1=0
+    c2=0
+    arate=0
+    srate=0
+  }else{
+    c1=cfg$migrations$c1
+    c2=cfg$migrations$c2
+    arate=cfg$migrations$arate
+    srate=cfg$migrations$srate
+    nmigration=nrow(cfg$migrations)
+  }
+  params[["nmigration"]]=nmigration
+  if(any(is.na(match(c(c1,c2),compartment$val)))){
+    stop("Can only specify migration for existing compartments")
+  }
   
   if(b_verbose){
     cat("MAX_EVENTS=",MAX_EVENTS,"\n")
     cat("MAX_SIZE=",MAX_SIZE,"\n")
-    ## browser()
   }
-  #bverbose=ifelse(b_verbose,1,0)
   res=.C("sim_pop2",
          edges=as.integer(edges),
          ndivs=as.integer(ndivs),
@@ -193,6 +210,7 @@ sim_pop=function(tree,
          nevents=as.integer(nevents),
          compartmentval=as.integer(compartment$val),
          compartmentrate=as.double(compartment$rate),
+         compartmentdeathrate=as.double(compartment$death_rate),
          compartmentsize=as.integer(compartment$popsize),
          ncompartment=as.integer(length(unique(compartment$val))),
          compinfoval=as.integer(compartmentinfo$val),
@@ -200,7 +218,7 @@ sim_pop=function(tree,
          compinfodriverstatus=as.integer(compartmentinfo$id),
          ncomp=as.integer(length(compartmentinfo$val)),
          params=as.double(params),
-         nparams=as.integer(length(params)),
+         #nparams=as.integer(length(params)),
          fitnessDistro = as.double(driversFitness),## ADD IN MAX_FITNESS_SIZE
          nMaxPreviousDriverID = as.integer(max_initial_driverid),
          max_size=as.integer(MAX_SIZE),
@@ -208,8 +226,13 @@ sim_pop=function(tree,
          trajectoryTs = as.double(trajectory_ts),
          trajectoryPop = as.integer(trajectory_pop),
          trajectoryDivRate = as.double(trajectory_div_rate),
+         trajectoryDeathRate = as.double(trajectory_death_rate),
          trajectoryCompartment = as.integer(trajectory_compartment),
          trajectorySize = as.integer(trajectory_size),
+         c1=as.integer(c1),
+         c2=as.integer(c2),
+         arate=as.double(arate),
+         srate=as.double(srate),
          driverSize=as.integer(length(driversFitness)),
          bVerbose=as.integer(b_verbose),
          edgesOut=integer(2*MAX_SIZE),
@@ -279,6 +302,9 @@ sim_pop=function(tree,
            ntips=res$ntipsOut,
            timestamp=res$eventTimestampOut[1:res$nEventsCount],
            pop.size=res$nPopSizeOut[1:res$nEventsCount],
+           pop.size.compartment=do.call("rbind",lapply(1:length(cfg$compartment$val),
+                                                       function(comp) data.frame(ts=res$eventTimestampOut[1:res$nEventsCount],val=comp-1,
+                                                                                 pop.size=res$nPopSizeOut[res$nEventsCount*comp+(1:res$nEventsCount)]))),
            totaldrivercount=res$nDriverOut[1:res$nEventsCount],
            ncellswithdrivers=1,
            params=params,
@@ -290,7 +316,9 @@ sim_pop=function(tree,
            driversPop = res$nAddedCompPops[1:res$nMaxDriverIDOut],
            maxDriverID = res$nMaxDriverIDOut
   )
-  
+  res$trajectory=res$pop.size.compartment %>% 
+    mutate(val=sprintf("c%d",val)) %>% pivot_wider(names_from=val,values_from = pop.size) %>% 
+    (function(x){for(col in names(x)[grep("^c",names(x))]){x[[col]]=ifelse(is.na(x[[col]]),0,x[[col]])};x})
   class(res)="simpop"
   res=get_tree_from_simpop(res,b_check)
   res$is_combined=FALSE
@@ -303,7 +331,17 @@ sim_pop=function(tree,
   res$is_combined=TRUE
   res
 }
-
+#' Plot simpop compartment level population trajectory
+#' @param simpop integer - a negative number uses current time to 100th second resoultion.
+#'
+#' @return ggplot
+#' @export
+plot_compartment_trajectory=function(simpop){
+  df=simpop$trajectory %>% pivot_longer(cols=names(simpop$trajectory)[grep("^c",names(simpop$trajectory))],names_to = "compartment",values_to = "pop")
+  df=df %>% filter(compartment != "c0") %>% 
+    left_join(simpop$cfg$compartment %>% mutate(compartment=sprintf("c%s",val))) %>% dplyr::select(ts,pop,desc)
+  ggplot(df,aes(x=ts,y=pop,col=desc))+geom_line()+theme_bw()+xlab("Timestamp")+ylab("Population")
+}
 
 get_max_popsize_from_trajectory=function(tree,trajectory){
   trajectory_ts = trajectory$ts; 
@@ -407,10 +445,10 @@ C_subsample_pop=function(tree,tips,trajectory=NULL){
   
   if(is.null(trajectory)){
     trajectory_ts = 0; trajectory_pop = 0
-    trajectory_div_rate = 0; trajectory_size=0; trajectory_compartment=1
+    trajectory_div_rate = 0; trajectory_death_rate = 0; trajectory_size=0; trajectory_compartment=1
   }else if(is.data.frame(trajectory) | is.data.table(trajectory)){
     trajectory_ts = trajectory$ts; trajectory_pop = trajectory$target_pop_size  
-    trajectory_div_rate = trajectory$division_rate; trajectory_size = nrow(trajectory)
+    trajectory_div_rate = trajectory$division_rate; trajectory_death_rate = trajectory$death_rate;trajectory_size = nrow(trajectory)
     trajectory_compartment = trajectory$compartment
   }
   
@@ -451,6 +489,7 @@ C_subsample_pop=function(tree,tips,trajectory=NULL){
          nevents=as.integer(nevents),
          compartmentval=as.integer(cfg$compartment$val),
          compartmentrate=as.double(cfg$compartment$rate),
+         compartmentdeathrate=as.double(0),
          compartmentsize=as.integer(cfg$compartment$popsize),
          ncompartment=as.integer(length(unique(cfg$compartment$val))),###
          compinfoval=as.integer(cfg$info$val),
@@ -465,6 +504,7 @@ C_subsample_pop=function(tree,tips,trajectory=NULL){
          trajectoryTs = as.double(trajectory_ts),
          trajectoryPop = as.double(trajectory_pop),
          trajectoryDivRate = as.double(trajectory_div_rate),
+         trajectoryDeathRate = as.double(trajectory_death_rate),
          trajectoryCompartment = as.integer(trajectory_compartment),
          trajectorySize = as.integer(trajectory_size),
          driversFitnessSize=as.integer(length(driversFitness)),
@@ -516,6 +556,7 @@ C_subsample_pop=function(tree,tips,trajectory=NULL){
            ntips=res$ntipsOut,
            timestamp=tree$timestamp,
            pop.size=tree$pop.size,
+           pop.size.compartment=tree$pop.size.compartment,
            totaldrivercount=tree$totaldrivercount,
            ncellswithdrivers=1,
            params=tree$params,
@@ -525,6 +566,8 @@ C_subsample_pop=function(tree,tips,trajectory=NULL){
            maxDriverID=tree$maxDriverID,
            status=res$status
   )
+  
+  res
 }
 
 
@@ -649,7 +692,11 @@ combine_simpops=function(simpop1,simpop2){
     ##cat("combine_simpops:Remember to add compartment level granularity")
     simpop2$timestamp=c(simpop1$timestamp,simpop2$timestamp)
     simpop2$pop.size=c(simpop1$pop.size,simpop2$pop.size)
+    simpop2$pop.size.compartment=rbind(simpop1$pop.size.compartment,simpop2$pop.size.compartment)
     simpop2$totaldrivercount=c(simpop1$totaldrivercount,simpop2$totaldrivercount)
+    simpop2$trajectory=simpop2$pop.size.compartment %>% 
+      mutate(val=sprintf("c%d",val)) %>% pivot_wider(names_from=val,values_from = pop.size) %>% 
+      (function(x){for(col in names(x)[grep("^c",names(x))]){x[[col]]=ifelse(is.na(x[[col]]),0,x[[col]])};x})
     simpop2
   }else{
     warning("Deprecated use of combine_simpops")
@@ -668,9 +715,9 @@ combine_simpops=function(simpop1,simpop2){
 #' @param basefit float - base fitness redundant now.
 #' @return list config
 #' @export
-getDefaultConfig=function(target_pop_size,rate,
+getDefaultConfig=function(target_pop_size,rate,death_rate=0,
                           descr="cellType1",ndriver=1,
-                          basefit=0.1)
+                          basefit=0,migrations=data.frame(c1=integer(),c2=integer(),arate=numeric(),srate=numeric()))
 {
   if(!is.numeric(rate) | length(rate)>1)
     stop("per day division rate should be a numeric value.")
@@ -682,6 +729,7 @@ getDefaultConfig=function(target_pop_size,rate,
     stop("the target population size should be a scalars")
   compartment=data.frame(val=0,
                          rate=-1,
+                         death_rate=0,
                          popsize=1,
                          desc="outgroup")
   info=data.frame(val=0,population=0,fitness=0,id=0,driver1=0)
@@ -690,7 +738,10 @@ getDefaultConfig=function(target_pop_size,rate,
     stop("please provide scalars")
   }
   #addCellCompartment(cfg,target_pop_size,rate,ndriver,basefit,descr)
-  addCellCompartment(cfg,target_pop_size,rate,descr=descr)
+  cfg=addCellCompartment(cfg,target_pop_size,rate,descr=descr,death_rate=death_rate)
+  ## check migrations... 
+  cfg$migrations=migrations
+  cfg
 }
 
 #' Adds a new cell compartment to a simpop configuration
@@ -701,7 +752,7 @@ getDefaultConfig=function(target_pop_size,rate,
 #' @param descr character. Description
 #' @return list simpop config
 #' @export
-addCellCompartment=function(cfg,population,rate,ndriver=0,basefit=0,descr=NULL){
+addCellCompartment=function(cfg,population,rate,ndriver=0,basefit=0,descr=NULL,death_rate=0){
   val=max(cfg$compartment$val)+1
   if(is.null(descr)){
     descr=sprintf("cellType%d",val)
@@ -713,7 +764,7 @@ addCellCompartment=function(cfg,population,rate,ndriver=0,basefit=0,descr=NULL){
   row$id=0
   
   cfg$info=rbind(cfg$info,row)
-  cfg$compartment=rbind(cfg$compartment,data.frame(val=val,rate=rate,popsize=population,desc=descr))
+  cfg$compartment=rbind(cfg$compartment,data.frame(val=val,rate=rate,popsize=population,desc=descr,death_rate=death_rate))
   ## TODO change this so adding empty works - changing this doesn't currently play nice with addDriverEvent
   cfg$drivers=rbind(cfg$drivers,data.frame(val=val,driver=1,fitness=0))#data.frame(val=integer(0),driver=integer(0),fitness=numeric(0)))#
   cfg
@@ -876,5 +927,21 @@ recalculateFitnessV2=function(cfg,cellType){
   }
   cfg
 }
-
-
+#' Gets subsampled tree - retaining events information appropriately
+#' @param tree simpop
+#' @param compartments integer. compartment ids (see val in tree$cfg)
+#' @param counts  integer. number of cells to sample from corresponding compartment.
+#' @return simpop
+#' @export
+#' @examples
+get_subsampled_tree_fixed_compartmentcount=function(tree,compartments,counts){
+ tipidx=which(tree$state==0 & tree$edge[,2]<=length(tree$tip.label))
+ for(i in 1:length(compartments)){
+   idx=which(tree$state==compartments[i] & tree$edge[,2]<=length(tree$tip.label))
+   if(length(idx)<counts[i]){
+     stop(sprintf("Attempt to sample more tips than available in compartment %d",compartments[i]))
+   }
+   tipidx=c(tipidx,sample(idx,counts[i]))
+ }
+ get_subsampled_tree(tree,-1,tree$edge[tipidx,2])
+}